requirements.txt
openai
python-docx
pywin32
python-dotenv
